Reworked the interface to the core functionality.
Due to the commit 240e0dbacd in the PG core, it's no longer possible to use
the original partition_table() function because this way the executor
registers a snapshot automatically, which causes an ERROR. Therefore, a
background worker is introduced here which calls the corresponding C function
directly. The partition_table() now only submits a "task" via shared memory
and waits until the worker is done with the work.